sample-fastapi/pyproject.toml
[project]
name = "sample-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.8",
    "python-jose>=3.3.0",
    "strawberry-graphql>=0.258.0",
    "uvicorn>=0.34.0",
]
